feat: move index page to root of the project
